refactor: trim fee-entrypoint smoke to negative-path only against pre-deployed services
