--- .github/workflows/spec-fee-entrypoint-smoke.yml
@@ -1,137 +1,0 @@
-name: Spec fee-entrypoint smoke
-
-on:
-  push:
-    branches:
-      - main
-    tags:
-      - "**"
-    paths:
-      - "contracts/**"
-      - "scripts/**"
-      - "services/**"
-      - "vendor/aztec-standards/**"
-      - "package.json"
-      - "bun.lock"
-      - "tsconfig*.json"
-      - ".aztecrc"
-      - ".github/workflows/spec-fee-entrypoint-smoke.yml"
-  pull_request:
-    types: [opened, synchronize, reopened, ready_for_review]
-  workflow_dispatch:
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  changes:
-    runs-on: ubuntu-latest
-    outputs:
-      relevant: ${{ steps.filter.outputs.relevant }}
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v6
-      - name: Detect relevant changes
-        id: filter
-        uses: dorny/paths-filter@v3
-        with:
-          filters: |
-            relevant:
-              - "contracts/**"
-              - "scripts/**"
-              - "services/**"
-              - "vendor/aztec-standards/**"
-              - "package.json"
-              - "bun.lock"
-              - "tsconfig*.json"
-              - ".aztecrc"
-              - ".github/workflows/spec-fee-entrypoint-smoke.yml"
-
-  fee-entrypoint-smoke:
-    needs: [changes]
-    if: github.event_name != 'pull_request' || needs.changes.outputs.relevant == 'true'
-    runs-on: ubuntu-latest
-    timeout-minutes: 60
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v6
-        with:
-          submodules: recursive
-
-      - name: Setup Node
-        uses: actions/setup-node@v6
-        with:
-          node-version: 24
-
-      - name: Setup Bun
-        uses: oven-sh/setup-bun@v2
-        with:
-          bun-version: "1.3.9"
-
-      - name: Install Foundry
-        uses: foundry-rs/foundry-toolchain@v1
-
-      - name: Read Aztec version
-        id: aztec-version
-        run: echo "version=$(cat .aztecrc | tr -d '\n')" >> $GITHUB_OUTPUT
-
-      - name: Add Aztec to PATH
-        run: |
-          echo "$HOME/.aztec/current/bin" >> $GITHUB_PATH
-          echo "$HOME/.aztec/current/node_modules/.bin" >> $GITHUB_PATH
-          echo "$HOME/.aztec/bin" >> $GITHUB_PATH
-
-      - name: Cache Aztec toolchain
-        uses: actions/cache@v5
-        with:
-          path: ~/.aztec
-          key: aztec-${{ runner.os }}-${{ steps.aztec-version.outputs.version }}
-
-      - name: Install Aztec
-        env:
-          VERSION: ${{ steps.aztec-version.outputs.version }}
-        run: |
-          output=$(aztec-up use 2>&1) || true
-          if echo "$output" | grep -q "Using aztec version $VERSION"; then
-            echo "Already using Aztec version $VERSION, skipping install"
-            exit 0
-          fi
-          bash -i <(curl -sL https://install.aztec.network/$VERSION)
-
-      - name: Install dependencies
-        run: bun install --frozen-lockfile
-
-      - name: Run fee-entrypoint smoke flow
-        env:
-          FPC_SMOKE_RESET_LOCAL_STATE: "1"
-        run: |
-          set -euo pipefail
-
-          aztec start --local-network >/tmp/aztec-local-network.log 2>&1 &
-          AZTEC_PID=$!
-
-          cleanup() {
-            if kill -0 "$AZTEC_PID" >/dev/null 2>&1; then
-              kill "$AZTEC_PID" >/dev/null 2>&1 || true
-              wait "$AZTEC_PID" >/dev/null 2>&1 || true
-            fi
-          }
-          trap cleanup EXIT INT TERM
-
-          for i in $(seq 1 180); do
-            if (echo >/dev/tcp/127.0.0.1/8080) >/dev/null 2>&1 && \
-               (echo >/dev/tcp/127.0.0.1/8545) >/dev/null 2>&1; then
-              echo "Aztec local network is reachable"
-              break
-            fi
-
-            if [[ "$i" == "180" ]]; then
-              echo "ERROR: local network did not become reachable in time" >&2
-              tail -n 200 /tmp/aztec-local-network.log >&2 || true
-              exit 1
-            fi
-            sleep 1
-          done
-
-          bun run smoke:fee-entrypoint:local
